move pagination hook from common to hooks
